feat: create group of nodes to test their presence on a workflow
